feat: add one template for info popup
