+++ frontend/src/components/FormRegistration/FormRegistration.jsx
@@ -1,10 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { useForm, Controller } from 'react-hook-form';
 import { InputText } from 'primereact/inputtext';
 import { Button } from 'primereact/button';
 import { Password } from 'primereact/password';
-import { Dialog } from 'primereact/dialog';
 import { Divider } from 'primereact/divider';
 import { classNames as cn } from 'primereact/utils';
 import styles from './FormRegistration.module.css';
@@ -12,11 +11,8 @@ import styles from './FormRegistration.module.css';
 export function FormRegistration({
 	showFormLogin,
 	handleRegister,
-	message,
-	isError,
+  setShowMessage,
 }) {
-	const [showMessage, setShowMessage] = useState(false);
-	const [formData, setFormData] = useState({});
 	const defaultValues = {
 		email: '',
 		password: '',
@@ -32,7 +28,6 @@ export function FormRegistration({
 	} = useForm({ defaultValues, mode: 'onChange' });
 
 	const onSubmit = (data) => {
-		setFormData(data);
 		handleRegister(data).then(() => setShowMessage(true));
 		reset();
 	};
@@ -40,16 +35,6 @@ export function FormRegistration({
 	const getFormErrorMessage = (name) =>
 		errors[name] && <small className="p-error">{errors[name].message}</small>;
 
-	const dialogFooter = (
-		<div className="flex justify-content-center">
-			<Button
-				label="OK"
-				className="p-button-text"
-				autoFocus
-				onClick={() => setShowMessage(false)}
-			/>
-		</div>
-	);
 	const passwordHeader = <h4>Введите пароль</h4>;
 	const passwordFooter = (
 		<>
@@ -64,48 +49,8 @@ export function FormRegistration({
 		</>
 	);
 
-	function handleDialog() {
-		if (isError) {
-			return (
-				<div className="flex justify-content-center flex-column pt-6 px-3">
-					<i
-						className="pi pi-times-circle"
-						style={{ fontSize: '5rem', color: 'var(--red-500)' }}
-					/>
-					<h4>Произошла ошибка!</h4>
-					<p style={{ lineHeight: 1.5 }}>{message}</p>
-				</div>
-			);
-		}
-
-		return (
-			<div className="flex justify-content-center flex-column pt-6 px-3">
-				<i
-					className="pi pi-check-circle"
-					style={{ fontSize: '5rem', color: 'var(--green-500)' }}
-				/>
-				<h4>Поздравляем!</h4>
-				<p style={{ lineHeight: 1.5 }}>
-					Вы зарегистрировались c Email <b>{formData.email}</b>
-				</p>
-			</div>
-		);
-	}
-
 	return (
 		<div className={styles.paddings}>
-			<Dialog
-				visible={showMessage}
-				onHide={() => setShowMessage(false)}
-				position="top"
-				footer={isError ? dialogFooter : ''}
-				showHeader={false}
-				breakpoints={{ '960px': '80vw' }}
-				style={{ width: '30vw' }}
-			>
-				<>{handleDialog()}</>
-			</Dialog>
-
 			<div className="flex justify-content-center">
 				<div className={styles.card}>
 					<h2 className="text-center">Регистрация</h2>
@@ -242,6 +187,5 @@ export function FormRegistration({
 FormRegistration.propTypes = {
 	showFormLogin: PropTypes.func.isRequired,
 	handleRegister: PropTypes.func.isRequired,
-	message: PropTypes.string.isRequired,
-	isError: PropTypes.bool.isRequired,
+	setShowMessage: PropTypes.func.isRequired,
 };
