+++ frontend/src/components/PopupLogin/PopupLogin.jsx
@@ -9,6 +9,7 @@ export function PopupLogin({
 	setVisible,
 	handleRegister,
 	handleLogin,
+  setShowMessage,
 	message,
 	isError,
 }) {
@@ -33,6 +34,7 @@ export function PopupLogin({
 					handleLogin={handleLogin}
 					message={message}
 					isError={isError}
+          setShowMessage={setShowMessage}
 				/>
 			) : (
 				<FormRegistration
@@ -40,6 +42,7 @@ export function PopupLogin({
 					handleRegister={handleRegister}
 					message={message}
 					isError={isError}
+          setShowMessage={setShowMessage}
 				/>
 			)}
 		</Dialog>
@@ -51,6 +54,7 @@ PopupLogin.propTypes = {
 	setVisible: PropTypes.func.isRequired,
 	handleRegister: PropTypes.func.isRequired,
 	handleLogin: PropTypes.func.isRequired,
+  setShowMessage: PropTypes.func.isRequired,
 	message: PropTypes.string.isRequired,
 	isError: PropTypes.bool.isRequired,
 };
